+++ src/main/java/com/roomify/shared/exception/user/RoleAlreadyAssignedException.java
@@ -1,0 +1,10 @@
+package com.roomify.shared.exception.user;
+
+import lombok.Builder;
+
+public class RoleAlreadyAssignedException extends Exception {
+    @Builder
+    public RoleAlreadyAssignedException(String message) {
+        super(message);
+    }
+}